+++ src/Rsse.Data/Domain/Dto/NoteRequestDto.cs
@@ -3,28 +3,16 @@ using System.Collections.Generic;
 namespace SearchEngine.Domain.Dto;
 
 /// <summary>
-/// Шаблон передачи данных для заметки
+/// Контейнер для запроса с заметкой.
 /// </summary>
+/// <param name="TagsCheckedRequest">Список отмеченных тегов.</param>
+/// <param name="TitleRequest">Именование заметки.</param>
+/// <param name="TextRequest">Текст заметки.</param>
+/// <param name="NoteIdExchange">Идентификатор сохраненной/измененной заметки, в обе стороны.</param>
 public record NoteRequestDto
-{
-    /// <summary>
-    /// Список отмеченных тегов в запросе
-    /// </summary>
-    public List<int>? TagsCheckedRequest { get; set; }
-
-    /// <summary>
-    /// Именование заметки в запросе
-    /// </summary>
-    public string? TitleRequest { get; set; }
-
-    /// <summary>
-    /// Текст заметки в запросе
-    /// </summary>
-    // ReSharper disable once PropertyCanBeMadeInitOnly.Global
-    public string? TextRequest { get; set; }
-
-    /// <summary>
-    /// Поле для передачи идентификатора сохраненной/измененной заметки в обе стороны
-    /// </summary>
-    public int NoteIdExchange { get; set; }
-}
+(
+    List<int>? TagsCheckedRequest,
+    string? TitleRequest,
+    string? TextRequest,
+    int NoteIdExchange
+);